+++ src/OpenAuthenticode.Module/AzureKeyVault.cs
@@ -6,7 +6,7 @@ using Azure.Security.KeyVault.Certificates;
 using Azure.Security.KeyVault.Keys.Cryptography;
 using AzureSignatureAlgorithm = Azure.Security.KeyVault.Keys.Cryptography.SignatureAlgorithm;
 
-namespace OpenAuthenticode.Shared;
+namespace OpenAuthenticode.Module;
 
 public sealed class AzureKey : KeyProvider, IDisposable
 {
@@ -99,65 +99,20 @@ public sealed class AzureKey : KeyProvider, IDisposable
     ~AzureKey() => Dispose();
 }
 
-internal sealed class AzureKeyVaultRSAKey : RSA
+internal sealed class AzureKeyVaultRSAKey : AzureRSAKey
 {
     private readonly CryptographyClient _client;
-
-    private readonly static byte[] _rsaSha1Digest = new byte[] {
-        0x30, 0x21, 0x30, 0x09, 0x06, 0x05, 0x2B, 0x0E,
-        0x03, 0x02, 0x1A, 0x05, 0x00, 0x04, 0x14,
-    };
 
     public AzureKeyVaultRSAKey(CryptographyClient client)
     {
         _client = client;
     }
 
-    public override byte[] SignHash(byte[] hash, HashAlgorithmName hashAlgorithm, RSASignaturePadding padding)
+    public override byte[] SignHashCore(byte[] hash, string azureAlgorithm)
     {
-        if (padding.Mode != RSASignaturePaddingMode.Pkcs1)
-        {
-            throw new CryptographicException($"Unsupported padding mode {padding.Mode}");
-        }
 
-        AzureSignatureAlgorithm sigAlgo;
-        if (hashAlgorithm == HashAlgorithmName.SHA1)
-        {
-            hash = CreateRSASha1Digest(hash);
-            sigAlgo = new AzureSignatureAlgorithm("RSNULL");
-        }
-        else if (hashAlgorithm == HashAlgorithmName.SHA256)
-        {
-            sigAlgo = AzureSignatureAlgorithm.RS256;
-        }
-        else if (hashAlgorithm == HashAlgorithmName.SHA384)
-        {
-            sigAlgo = AzureSignatureAlgorithm.RS384;
-        }
-        else if (hashAlgorithm == HashAlgorithmName.SHA512)
-        {
-            sigAlgo = AzureSignatureAlgorithm.RS512;
-        }
-        else
-        {
-            string msg = "Support for the hash algorithm requested '{0}' for this RSA key has not been implemented";
-            throw new CryptographicException(string.Format(msg, hashAlgorithm.Name));
-        }
-
+        AzureSignatureAlgorithm sigAlgo = new(azureAlgorithm);
         return _client.Sign(sigAlgo, hash).Signature;
-    }
-
-    public override RSAParameters ExportParameters(bool includePrivateParameters) => throw new NotImplementedException();
-
-    public override void ImportParameters(RSAParameters parameters) => throw new NotImplementedException();
-
-    private static byte[] CreateRSASha1Digest(byte[] hash)
-    {
-        byte[] pkcs1Digest = new byte[_rsaSha1Digest.Length + 20];
-        _rsaSha1Digest.CopyTo(pkcs1Digest, 0);
-        hash.CopyTo(pkcs1Digest, _rsaSha1Digest.Length);
-
-        return pkcs1Digest;
     }
 }